Divide - Stub step - tests passed. Change type input and output to double to support the double results. In refactor I adding support the double input in all methods.

+++ Calculator.Tests/CalculatorTests.cs
@@ -60,7 +60,7 @@ public class CalculatorTests
     [DataRow(10, 10, 1)]
     [DataRow(25, 10, 2.5)]
     [DataRow(10, 25, 0.4)]
-    public void GivenTwoPositiveNumbers_WhenDivide_ThenReturnsCorrectValue(int a, int b, int expected)
+    public void GivenTwoPositiveNumbers_WhenDivide_ThenReturnsCorrectValue(double a, double b, double expected)
     {
         var result = Calculator.Divide(a, b);
         Assert.AreEqual(result, expected);
@@ -70,7 +70,7 @@ public class CalculatorTests
     [TestMethod]
     [DataRow(-120, -3, 40)]
     [DataRow(-120, 3, -40)]
-    public void GivenTwoNegativeNumbers_WhenDivide_ThenReturnsCorrectValue(int a, int b, int expected)
+    public void GivenTwoNegativeNumbers_WhenDivide_ThenReturnsCorrectValue(double a, double b, double expected)
     {
         var result = Calculator.Divide(a, b);
         Assert.AreEqual(result, expected);
